server/requirements.txt
langgraph
langchain
langchain-groq
langchain-community
psycopg2-binary
faiss-cpu
python-dotenv
PyPDF2